+++ Web/Scotty/Action.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PackageImports #-}
 {-# LANGUAGE CPP               #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RankNTypes        #-}
@@ -58,11 +61,13 @@ import           Control.Monad.Trans.Except
 import           Control.Concurrent.MVar
 
 import qualified Data.Aeson                 as A
+import Data.Bool (bool)
 import qualified Data.ByteString.Char8      as B
 import qualified Data.ByteString.Lazy.Char8 as BL
 import qualified Data.CaseInsensitive       as CI
 import           Data.Default.Class         (def)
 import           Data.Int
+import           Data.String (IsString(..))
 import qualified Data.Text                  as ST
 import qualified Data.Text.Encoding         as STE
 import qualified Data.Text.Lazy             as T
@@ -80,51 +85,74 @@ import           Network.Wai (Request, Response, StreamingBody, Application, req
 import           Numeric.Natural
 
 import           Prelude ()
-import           Prelude.Compat
+import "base-compat-batteries" Prelude.Compat
 
 import           Web.Scotty.Internal.Types
-import           Web.Scotty.Util (mkResponse, addIfNotPresent, add, replace, lazyTextToStrictByteString, strictByteStringToLazyText, catch, catchAny, tryAny)
+import           Web.Scotty.Util (mkResponse, addIfNotPresent, add, replace, lazyTextToStrictByteString, strictByteStringToLazyText, catch, catches, catchesOptionally, catchAny, try, tryAny)
 
 import Network.Wai.Internal (ResponseReceived(..))
 
--- Nothing indicates route failed (due to Next) and pattern matching should continue.
+-- | Nothing indicates route failed (due to Next) and pattern matching should continue.
 -- Just indicates a successful response.
 -- runAction :: (ScottyError e, Monad m) => ErrorHandler e m -> ActionEnv -> ActionT e m () -> m (Maybe Response)
--- runAction :: MonadUnliftIO m =>
---              Maybe a -> ActionEnv -> ActionT e m () -> m (Maybe Response)
 runAction :: MonadUnliftIO m =>
-             Maybe (AErr -> ActionT e m ())
-          -> ActionEnv -> ActionT e m () -> m (Maybe Response)
-runAction h env action = do
-  ei <- flip runReaderT env $ runAM $ tryAny (action `catch` defH h)
+             Maybe (ErrorHandler m) -- ^ if present, this handler is in charge of user-defined exceptions
+          -> ActionEnv -> ActionT m () -> m (Maybe Response)
+runAction mh env action = do
+  ok <- flip runReaderT env $ runAM $ tryNext (catchesOptionally action mh actionErrorHandler )
   res <- getResponse env
-  return $ either (const Nothing) (const $ Just $ mkResponse res) ei
+  return $ bool Nothing (Just $ mkResponse res) ok
+  -- return $ either (const Nothing) (const $ Just $ mkResponse res) ei
 
 
--- | Default error handler for all actions.
--- defH :: (Monad m) => ErrorHandler e m -> ActionError -> ActionT e m ()
--- defH :: MonadIO m => Maybe a -> ActionError -> ActionT e m ()
-defH :: MonadUnliftIO m => Maybe (AErr -> ActionT e m ()) -> ActionError -> ActionT e m ()
-defH _          (Redirect url)    = do
+{-|
+* ActionError should only be caught by runAction
+* handlers should not throw ActionError (= do not export ActionError constructors)
+-}
+
+tryNext :: MonadUnliftIO m => m a -> m Bool
+tryNext io = catch (io >> pure True) $ \e ->
+  case e of
+    Next -> pure False
+    _ -> pure True
+
+-- | Error handler in charge of 'ActionError'. Rethrowing 'Next' here is caught by 'tryNext'
+actionErrorHandler :: MonadIO m => ErrorHandler m -- ActionError -> ActionT m ()
+actionErrorHandler = Handler $ \case
+  Redirect url -> do
     status status302
     setHeader "Location" url
-defH Nothing    (ActionError s e)   = do
+  StatusError s e -> do
     status s
     let code = T.pack $ show $ statusCode s
     let msg = T.fromStrict $ STE.decodeUtf8 $ statusMessage s
-    html $ mconcat ["<h1>", code, " ", msg, "</h1>", T.pack (show e)]
-defH h@(Just f) (ActionError _ e)   = f e `catch` (defH h) -- so handlers can throw exceptions themselves -- TODO
-defH _          Next              = next
-defH _          Finish            = return ()
+    html $ mconcat ["<h1>", code, " ", msg, "</h1>", e]
+  Next -> next
+  Finish -> return ()
+
+-- -- defH :: (Monad m) => ErrorHandler e m -> ActionError -> ActionT e m ()
+-- defH :: MonadUnliftIO m => Maybe (AErr -> ActionT m ()) -> ActionError -> ActionT m ()
+-- defH _          (Redirect url)    = do
+--     status status302
+--     setHeader "Location" url
+-- defH Nothing    (ActionError s e)   = do
+--     status s
+--     let code = T.pack $ show $ statusCode s
+--     let msg = T.fromStrict $ STE.decodeUtf8 $ statusMessage s
+--     html $ mconcat ["<h1>", code, " ", msg, "</h1>", T.pack (show e)]
+-- defH h@(Just f) (ActionError _ e)   = f e `catch` (defH h) -- so handlers can throw exceptions themselves -- TODO
+-- defH _          Next              = next      -- rethrow 'Next'
+-- defH _          Finish            = return () -- stop
 
 -- | Throw an exception, which can be caught with 'rescue'. Uncaught exceptions
 -- turn into HTTP 500 responses.
-raise :: (Monad m) => AErr -> ActionT e m a
-raise = raiseStatus status500
+-- raise :: (Monad m) => AErr -> ActionT m a
+raise :: (MonadIO m, E.Exception e) => e -> ActionT m a
+raise = E.throw -- raiseStatus status500
 
 -- | Throw an exception, which can be caught with 'rescue'. Uncaught exceptions turn into HTTP responses corresponding to the given status.
-raiseStatus :: (Monad m) => Status -> AErr -> ActionT e m a
-raiseStatus s = E.throw . ActionError s
+raiseStatus :: (Monad m) => Status -> T.Text -> ActionT m a
+raiseStatus s = E.throw . StatusError s
 
 
 -- | Abort execution of this action and continue pattern matching routes.
@@ -141,23 +169,21 @@ raiseStatus s = E.throw . ActionError s
 -- > get "/foo/:baz" $ do
 -- >   w <- param "baz"
 -- >   text $ "You made a request to: " <> w
-next :: (Monad m) => ActionT e m a
+next :: (Monad m) => ActionT m a
 next = E.throw Next
 
 -- | Catch an exception thrown by 'raise'.
 --
--- > raise "just kidding" `rescue` (\msg -> text msg)
-rescue :: (MonadUnliftIO m) => ActionT e m a -> (AErr -> ActionT e m a) -> ActionT e m a
-rescue action h = catch action $ \e -> case e of
-  ActionError _ err -> h err -- handle errors
-  other -> E.throw other -- rethrow internal error types
+-- > raise JustKidding `rescue` (\msg -> text msg)
+rescue :: (MonadUnliftIO m, E.Exception e) => ActionT m a -> (e -> ActionT m a) -> ActionT m a
+rescue = catch
 
--- | Like 'liftIO', but catch any IO exceptions and turn them into 'ScottyError's.
-liftAndCatchIO = undefined -- FIXME
--- liftAndCatchIO :: (ScottyError e, MonadIO m) => IO a -> ActionT e m a
--- liftAndCatchIO io = ActionT $ do
---     r <- liftIO $ liftM Right io `E.catch` (\ e -> return $ Left $ stringError $ show (e :: E.SomeException))
---     either throwError return r
+-- | Catch any synchronous IO exceptions
+liftAndCatchIO :: MonadIO m => IO a -> ActionT m a
+liftAndCatchIO io = liftIO $ do
+  r <- tryAny io
+  either E.throwIO pure r
+
 
 -- | Redirect to given URL. Like throwing an uncatchable exception. Any code after the call to redirect
 -- will not be run.
@@ -167,7 +193,7 @@ liftAndCatchIO = undefined -- FIXME
 -- OR
 --
 -- > redirect "/foo/bar"
-redirect :: (Monad m) => T.Text -> ActionT e m a
+redirect :: (Monad m) => T.Text -> ActionT m a
 redirect = E.throw . Redirect
 
 -- | Finish the execution of the current action. Like throwing an uncatchable
@@ -175,25 +201,25 @@ redirect = E.throw . Redirect
 --
 -- /Since: 0.10.3/
 -- finish :: (ScottyError e, Monad m) => ActionT e m a
-finish :: (Monad m) => ActionT e m a
+finish :: (Monad m) => ActionT m a
 finish = E.throw Finish
 
 -- | Get the 'Request' object.
-request :: Monad m => ActionT e m Request
+request :: Monad m => ActionT m Request
 request = ActionT $ envReq <$> ask
 
 -- | Get list of uploaded files.
-files :: Monad m => ActionT e m [File]
+files :: Monad m => ActionT m [File]
 files = ActionT $ envFiles <$> ask
 
 -- | Get a request header. Header name is case-insensitive.
-header :: (Monad m) => T.Text -> ActionT e m (Maybe T.Text)
+header :: (Monad m) => T.Text -> ActionT m (Maybe T.Text)
 header k = do
     hs <- requestHeaders <$> request
     return $ fmap strictByteStringToLazyText $ lookup (CI.mk (lazyTextToStrictByteString k)) hs
 
 -- | Get all the request headers. Header names are case-insensitive.
-headers :: (Monad m) => ActionT e m [(T.Text, T.Text)]
+headers :: (Monad m) => ActionT m [(T.Text, T.Text)]
 headers = do
     hs <- requestHeaders <$> request
     return [ ( strictByteStringToLazyText (CI.original k)
@@ -201,13 +227,13 @@ headers = do
            | (k,v) <- hs ]
 
 -- | Get the request body.
-body :: (ScottyError e,  MonadIO m) => ActionT e m BL.ByteString
+body :: (MonadIO m) => ActionT m BL.ByteString
 body = ActionT ask >>= (liftIO . envBody)
 
 -- | Get an IO action that reads body chunks
 --
 -- * This is incompatible with 'body' since 'body' consumes all chunks.
-bodyReader :: Monad m => ActionT e m (IO B.ByteString)
+bodyReader :: Monad m => ActionT m (IO B.ByteString)
 bodyReader = ActionT $ envBodyChunk <$> ask
 
 -- | Parse the request body as a JSON object and return it.
@@ -219,27 +245,26 @@ bodyReader = ActionT $ envBodyChunk <$> ask
 --   422 Unprocessable Entity.
 --
 --   These status codes are as per https://www.restapitutorial.com/httpstatuscodes.html.
-jsonData = undefined
--- jsonData :: (A.FromJSON a, ScottyError e, MonadIO m) => ActionT e m a
--- jsonData = do
---     b <- body
---     when (b == "") $ do
---       let htmlError = "jsonData - No data was provided."
---       raiseStatus status400 $ stringError htmlError
---     case A.eitherDecode b of
---       Left err -> do
---         let htmlError = "jsonData - malformed."
---               `mappend` " Data was: " `mappend` BL.unpack b
---               `mappend` " Error was: " `mappend` err
---         raiseStatus status400 $ stringError htmlError
---       Right value -> case A.fromJSON value of
---         A.Error err -> do
---           let htmlError = "jsonData - failed parse."
---                 `mappend` " Data was: " `mappend` BL.unpack b `mappend` "."
---                 `mappend` " Error was: " `mappend` err
---           raiseStatus status422 $ stringError htmlError
---         A.Success a -> do
---           return a
+jsonData :: (A.FromJSON a, MonadIO m) => ActionT m a
+jsonData = do
+    b <- body
+    when (b == "") $ do
+      let htmlError = "jsonData - No data was provided."
+      raiseStatus status400 $ T.pack htmlError
+    case A.eitherDecode b of
+      Left err -> do
+        let htmlError = "jsonData - malformed."
+              `mappend` " Data was: " `mappend` BL.unpack b
+              `mappend` " Error was: " `mappend` err
+        raiseStatus status400 $ T.pack htmlError
+      Right value -> case A.fromJSON value of
+        A.Error err -> do
+          let htmlError = "jsonData - failed parse."
+                `mappend` " Data was: " `mappend` BL.unpack b `mappend` "."
+                `mappend` " Error was: " `mappend` err
+          raiseStatus status422 $ T.pack htmlError
+        A.Success a -> do
+          return a
 
 -- | Get a parameter. First looks in captures, then form data, then query parameters.
 --
@@ -248,11 +273,11 @@ jsonData = undefined
 -- * If parameter is found, but 'parseParam' fails to parse to the correct type, 'next' is called.
 --   This means captures are somewhat typed, in that a route won't match if a correctly typed
 --   capture cannot be parsed.
--- param :: (Parsable a, ScottyError e, Monad m) => T.Text -> ActionT e m a
+param :: (Parsable a, MonadIO m) => T.Text -> ActionT m a
 param k = do
     val <- ActionT $ (lookup k . getParams) <$> ask
     case val of
-        -- Nothing -> raise $ stringError $ "Param: " ++ T.unpack k ++ " not found!" -- FIXME
+        Nothing -> raiseStatus status500 $ "Param: " <> k <> " not found!" -- FIXME
         Just v  -> either (const next) return $ parseParam v
 {-# DEPRECATED param "(#204) Not a good idea to treat all parameters identically. Use captureParam, formParam and queryParam instead. "#-}
 
@@ -262,7 +287,7 @@ param k = do
 --
 -- * If the parameter is found, but 'parseParam' fails to parse to the correct type, 'next' is called.
 -- captureParam :: (Parsable a, ScottyError e, Monad m) => T.Text -> ActionT e m a
-captureParam :: (Parsable a, Monad m) => T.Text -> ActionT e m a
+captureParam :: (Parsable a, Monad m) => T.Text -> ActionT m a
 captureParam = paramWith CaptureParam envCaptureParams status500
 
 -- | Get a form parameter.
@@ -271,7 +296,7 @@ captureParam = paramWith CaptureParam envCaptureParams status500
 --
 -- * This function raises a code 400 also if the parameter is found, but 'parseParam' fails to parse to the correct type.
 -- formParam :: (Parsable a, ScottyError e, Monad m) => T.Text -> ActionT e m a
-formParam :: (Parsable a, Monad m) => T.Text -> ActionT e m a
+formParam :: (Parsable a, Monad m) => T.Text -> ActionT m a
 formParam = paramWith FormParam envFormParams status400
 
 -- | Get a query parameter.
@@ -280,7 +305,7 @@ formParam = paramWith FormParam envFormParams status400
 --
 -- * This function raises a code 400 also if the parameter is found, but 'parseParam' fails to parse to the correct type.
 -- queryParam :: (Parsable a, ScottyError e, Monad m) => T.Text -> ActionT e m a
-queryParam :: (Parsable a, Monad m) => T.Text -> ActionT e m a
+queryParam :: (Parsable a, Monad m) => T.Text -> ActionT m a
 queryParam = paramWith QueryParam envQueryParams status400
 
 data ParamType = CaptureParam
@@ -297,33 +322,33 @@ paramWith :: (Monad m, Parsable b) =>
           -> (ActionEnv -> [Param])
           -> Status -- ^ HTTP status to return if parameter is not found
           -> T.Text -- ^ parameter name
-          -> ActionT e m b
+          -> ActionT m b
 paramWith ty f err k = do
     val <- ActionT $ (lookup k . f) <$> ask
     case val of
-      -- Nothing -> raiseStatus err $ stringError (unwords [show ty, "parameter:", T.unpack k, "not found!"]) -- FIXME
+      Nothing -> raiseStatus err (T.unwords [T.pack (show ty), "parameter:", k, "not found!"])
       Just v ->
         let handleParseError = \case
               CaptureParam -> next
-              -- _ -> raiseStatus err $ stringError (unwords ["Cannot parse", T.unpack v, "as a", show ty, "parameter"]) -- FIXME
+              _ -> raiseStatus err (T.unwords ["Cannot parse", v, "as a", T.pack (show ty), "parameter"])
         in either (const $ handleParseError ty) return $ parseParam v
 
 -- | Get all parameters from capture, form and query (in that order).
-params :: Monad m => ActionT e m [Param]
+params :: Monad m => ActionT m [Param]
 params = paramsWith getParams
 {-# DEPRECATED params "(#204) Not a good idea to treat all parameters identically. Use captureParams, formParams and queryParams instead. "#-}
 
 -- | Get capture parameters
-captureParams :: Monad m => ActionT e m [Param]
+captureParams :: Monad m => ActionT m [Param]
 captureParams = paramsWith envCaptureParams
 -- | Get form parameters
-formParams :: Monad m => ActionT e m [Param]
+formParams :: Monad m => ActionT m [Param]
 formParams = paramsWith envFormParams
 -- | Get query parameters
-queryParams :: Monad m => ActionT e m [Param]
+queryParams :: Monad m => ActionT m [Param]
 queryParams = paramsWith envQueryParams
 
-paramsWith :: Monad m => (ActionEnv -> a) -> ActionT e m a
+paramsWith :: Monad m => (ActionEnv -> a) -> ActionT m a
 paramsWith f = ActionT (f <$> ask)
 
 {-# DEPRECATED getParams "(#204) Not a good idea to treat all parameters identically" #-}
@@ -394,35 +419,35 @@ readEither t = case [ x | (x,"") <- reads (T.unpack t) ] of
                 _   -> Left "readEither: ambiguous parse"
 
 -- | Set the HTTP response status. Default is 200.
-status :: Monad m => Status -> ActionT e m ()
-status = undefined -- FIXME
+status :: MonadIO m => Status -> ActionT m ()
+status = modifyResponse . setStatus
 
 -- Not exported, but useful in the functions below.
 changeHeader :: MonadIO m
              => (CI.CI B.ByteString -> B.ByteString -> [(HeaderName, B.ByteString)] -> [(HeaderName, B.ByteString)])
-             -> T.Text -> T.Text -> ActionT e m ()
+             -> T.Text -> T.Text -> ActionT m ()
 changeHeader f k =
   modifyResponse . setHeaderWith . f (CI.mk $ lazyTextToStrictByteString k) . lazyTextToStrictByteString
 
 -- | Add to the response headers. Header names are case-insensitive.
-addHeader :: MonadIO m => T.Text -> T.Text -> ActionT e m ()
+addHeader :: MonadIO m => T.Text -> T.Text -> ActionT m ()
 addHeader = changeHeader add
 
 -- | Set one of the response headers. Will override any previously set value for that header.
 -- Header names are case-insensitive.
-setHeader :: MonadIO m => T.Text -> T.Text -> ActionT e m ()
+setHeader :: MonadIO m => T.Text -> T.Text -> ActionT m ()
 setHeader = changeHeader replace
 
 -- | Set the body of the response to the given 'T.Text' value. Also sets \"Content-Type\"
 -- header to \"text/plain; charset=utf-8\" if it has not already been set.
-text :: (MonadIO m) => T.Text -> ActionT e m ()
+text :: (MonadIO m) => T.Text -> ActionT m ()
 text t = do
     changeHeader addIfNotPresent "Content-Type" "text/plain; charset=utf-8"
     raw $ encodeUtf8 t
 
 -- | Set the body of the response to the given 'T.Text' value. Also sets \"Content-Type\"
 -- header to \"text/html; charset=utf-8\" if it has not already been set.
-html :: (MonadIO m) => T.Text -> ActionT e m ()
+html :: (MonadIO m) => T.Text -> ActionT m ()
 html t = do
     changeHeader addIfNotPresent "Content-Type" "text/html; charset=utf-8"
     raw $ encodeUtf8 t
@@ -430,15 +455,15 @@ html t = do
 -- | Send a file as the response. Doesn't set the \"Content-Type\" header, so you probably
 -- want to do that on your own with 'setHeader'. Setting a status code will have no effect
 -- because Warp will overwrite that to 200 (see 'Network.Wai.Handler.Warp.Internal.sendResponse').
-file :: Monad m => FilePath -> ActionT e m ()
-file = undefined -- FIXME
+file :: MonadIO m => FilePath -> ActionT m ()
+file = modifyResponse . setContent . ContentFile
 
-rawResponse :: Monad m => Response -> ActionT e m ()
-rawResponse = undefined -- FIXME
+rawResponse :: MonadIO m => Response -> ActionT m ()
+rawResponse = modifyResponse . setContent . ContentResponse
 
 -- | Set the body of the response to the JSON encoding of the given value. Also sets \"Content-Type\"
 -- header to \"application/json; charset=utf-8\" if it has not already been set.
-json :: (A.ToJSON a, MonadIO m) => a -> ActionT e m ()
+json :: (A.ToJSON a, MonadIO m) => a -> ActionT m ()
 json v = do
     changeHeader addIfNotPresent "Content-Type" "application/json; charset=utf-8"
     raw $ A.encode v
@@ -446,18 +471,18 @@ json v = do
 -- | Set the body of the response to a Source. Doesn't set the
 -- \"Content-Type\" header, so you probably want to do that on your
 -- own with 'setHeader'.
-stream :: MonadIO m => StreamingBody -> ActionT e m ()
+stream :: MonadIO m => StreamingBody -> ActionT m ()
 stream = modifyResponse . setContent . ContentStream
 
 -- | Set the body of the response to the given 'BL.ByteString' value. Doesn't set the
 -- \"Content-Type\" header, so you probably want to do that on your
 -- own with 'setHeader'.
-raw :: MonadIO m => BL.ByteString -> ActionT e m ()
+raw :: MonadIO m => BL.ByteString -> ActionT m ()
 raw = modifyResponse . setContent . ContentBuilder . fromLazyByteString
 
 -- | Nest a whole WAI application inside a Scotty handler.
 -- See Web.Scotty for further documentation
-nested :: (MonadIO m) => Network.Wai.Application -> ActionT e m ()
+nested :: (MonadIO m) => Network.Wai.Application -> ActionT m ()
 nested app = do
   -- Is MVar really the best choice here? Not sure.
   r <- request